requirements.txt
annotated-types==0.7.0
anyio==4.9.0
anywidget==0.9.18
app==0.0.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
commcare-api==0.1.0a0
contourpy==1.3.1
cryptography==45.0.3
cycler==0.12.1
dash==3.0.3
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dotenv==0.9.9
easydev==0.13.3
et_xmlfile==2.0.0
executing==2.2.0
fabric==3.2.2
fake==0.8
Faker==37.3.0
fastjsonschema==2.21.1
Flask==3.0.3
folium==0.19.6
fonttools==4.56.0
fqdn==1.5.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.2.0
h11==0.16.0
htmltools==0.6.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
invoke==2.2.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itables==2.4.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaleido==0.2.1
kiwisolver==1.4.8
line_profiler==4.2.0
linkify-it-py==2.0.3
load-dotenv==0.1.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.3
mizani==0.13.3
mpld3==0.5.10
mysql-connector-python==9.2.0
narwhals==1.30.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.2.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.1.2
plotnine==0.14.5
prometheus_client==0.22.0
prompt_toolkit==3.0.50
psutil==7.0.0
psygnal==0.13.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyodata==1.11.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.1
pywin32==309
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.3.0
quarto==0.1.0
questionary==2.1.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
rpy2==3.5.17
rsa==4.9
rtools==0.2.12
scipy==1.15.2
seaborn==0.13.2
selenium==4.29.0
Send2Trash==1.8.3
setuptools==78.1.0
shiny==1.4.0
shinywidgets==0.6.2
six==1.17.0
sniffio==1.3.1
sns==0.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.47.0
statsmodels==0.14.4
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wrapt==1.17.2
wsproto==1.2.0
XlsxWriter==3.2.2
xyzservices==2025.4.0
zipp==3.21.0
